fix: duration in arabic format, episode details month not translated to arabic it's stay in english, episode details rate should be integer not double, Time in episode details not same figma in Arabic shown but in english is hidden

+++ entity/src/main/java/com/baghdad/entity/media/Episode.kt
@@ -6,7 +6,7 @@ data class Episode(
     val id: Long,
     val title: String,
     val episodeNumber: Int,
-    val rating: Double,
+    val rating: Int,
     val duration: String,
     val releasedDate: LocalDate?,
     val trailerUrl: String,
